Reworked ddi_node representation
Directly referencing the XML structure proved to be inflexible when assigning subcomponents to parent structures. Since xml2 automatically removes nodes when out of scope, the act of changing the parent would inadvertently mark the component as no longer referenced, creating a possible segfault situation. Now, ddi_nodes are represented as R lists and then converted to XML using as_xml(), which ensures that the parent-child relationship is safeguarded.